Separate screen sharing prompts from realtime transport
