+++ src/runtime/codex-realtime/screen-observation.test.ts
@@ -1,5 +1,5 @@
 import { afterEach, describe, expect, it, vi } from "vitest";
-import { ScreenObservationTransport, screenPointerSettingText } from "./screen-observation";
+import { ScreenObservationTransport } from "./screen-observation";
 
 const frame = {
   frameId: "frame-1",
@@ -52,16 +52,6 @@ describe("screen observation transport", () => {
       "the image is stale, or the target moved, inspect a fresh shared image before pointing again",
     );
     expect(text).toContain("No response is needed for the capture itself");
-  });
-
-  it("makes an ON setting useful without treating the setting change as a request", () => {
-    const text = screenPointerSettingText(true);
-    expect(text).toContain("While sharing is active, proactively use a marker");
-    expect(text).toContain("latest inspected shared image");
-    expect(text).toContain("no separate request to point is needed");
-    expect(text).toContain("Omit markers for unrelated conversation, uncertain targets");
-    expect(text).toContain("after an OFF/ON transition, inspect a newer shared image");
-    expect(text).toContain("not a request to act or speak");
   });
 
   it.each([
